core: remove unnecessary calls to __tls_init
__tls_init is a function generated by gcc and inserted prior to accesses of
thread local variables that might need to be initialized. The function checks
to see if the variable has been initialized already and if it has then it
returns immediately, otherwise it starts some (potentially expensive)
initialization work.

The trouble with this is that __tls_init is called unconditionally adding
massively to the overhead of thread local variables, particularly on systems
where the base overhead of a function call is high. Ideally the check should
be inlined into the calling function and a function call should only be made
if the variable is actually uninitialized.

This change works around this problem for some key thread local variables
accessed by allocate() and free() by lazily initializing them explicitly.