thread: make thread_context::s_main portable
On some platforms (e.g. IBM Z) the ABI requires that arguments are
sign/zero extended to register width as appropriate for their type.
This means that signed and unsigned arguments cannot be used
interchangeably since the compiler can assume that it doesn't have
to explicitly zero extend an unsigned argument or sign extend a
signed argument.

s_main is called indirectly and the caller assumes that it takes
signed ints, so s_main must accept signed int arguments. If it
expects unsigned ints then the thread_context pointer will not be
rematerialized correctly when lo&(1<<31) != 0.

Message-Id: <20170905234423.82554-1-mike.munday@ibm.com>